feat: add user repository
